+++ ProdManager/models/Maintenance.py
@@ -7,6 +7,7 @@ from ProdManager import db
 from ProdManager import lang
 from ProdManager.helpers.model import ModelEnum
 from ProdManager.helpers.links import custom_url_for
+import ProdManager.helpers.resource as ResourceHelpers
 
 from .Service import ServiceStatus
 from .MaintenanceEvent import MaintenanceEvent
@@ -21,7 +22,7 @@ class Maintenance(db.Model):
   id = Column(Integer, primary_key=True)
   name = Column(String(), nullable=False)
   description = Column(String(), nullable=True)
-  external_reference = Column(String(), nullable=True)
+  external_reference = Column(String(), nullable=True, index=True)
   status = Column(Enum(MaintenanceStatus), nullable=False, default=MaintenanceStatus.SCHEDULED)
   scope_id = Column(Integer, ForeignKey('scope.id'), nullable=False)
   service_id = Column(Integer, ForeignKey('service.id'), nullable=False)
@@ -138,5 +139,15 @@ class Maintenance(db.Model):
 
     if self.external_reference:
       result = f"[{self.external_reference}]" + result
+
+    return result
+
+  @classmethod
+  def count_by_status(cls, query, serialize=False, filters=()):
+    result = dict()
+
+    for status in MaintenanceStatus:
+      key = status.value if serialize else status
+      result[key] = ResourceHelpers.count_in_status_from_query(cls, query, status, filters)
 
     return result